pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "schroneko", email = "schroneko.fuller@gmail.com" }
]
dependencies = [
    "discord-py>=2.3.2",
    "anthropic>=0.21.3",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "ruff>=0.3.4",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/discordbot"]
